feat: Implementação do SignalR Inicio implementação testes unitários

+++ App.Desafio.Blog.Application/Services/BlogService.cs
@@ -1,29 +1,40 @@
-﻿using App.Desafio.Blog.Domain.Dtos.Requests;
+﻿using App.Desafio.Blog.Crosscutting.Sockets;
+using App.Desafio.Blog.Domain.Dtos.Requests;
 using App.Desafio.Blog.Domain.Dtos.Responses;
 using App.Desafio.Blog.Domain.Enums;
 using App.Desafio.Blog.Domain.Exceptions;
 using App.Desafio.Blog.Domain.Extensions;
 using App.Desafio.Blog.Domain.Interfaces;
+using Microsoft.AspNetCore.SignalR;
 
 namespace App.Desafio.Blog.Application.Services
 {
     public class BlogService : IBlogService
     {
         private readonly IBlogRepository _blogRepository;
-        public BlogService(IBlogRepository blogRepository)
+        private readonly IHubContext<BlogChannel> _hubContext;
+        public BlogService(IBlogRepository blogRepository, IHubContext<BlogChannel> hubContext)
         { 
             _blogRepository = blogRepository;
+            _hubContext = hubContext;
         }
         public async Task<PostResponse> CreatePostAsync(CreatePostRequest request, Guid userId)
         {
             var post = request.DtoToPost(userId);
             var response = await _blogRepository.CreatePostAsync(post);
+
+            await _hubContext.Clients.All.SendAsync("ReceiveMessage", "A new post has been created!");
+
             return post.DtoToPost();
         }
 
         public async Task DeletePostAsync(Guid id, Guid userId)
         {
             var post = await _blogRepository.GetPostByUserIdAndPostId(id, userId);
+
+            if (post != null) {
+                throw new DomainException(DomainErrorCode.NotFound, "Post not found."); 
+            }
             
             await _blogRepository.DeletePostAsync(post);
         }
@@ -61,9 +72,14 @@ namespace App.Desafio.Blog.Application.Services
 
         public async Task<PostResponse> UpdatePostAsync(UpdatePostRequest request, Guid userId)
         {
-            await _blogRepository.GetPostByUserIdAndPostId(request.PostId, userId);
+            var post = await _blogRepository.GetPostByUserIdAndPostId(request.PostId, userId);
 
-            var post = request.DtoToPost(userId);
+            if (post is null)
+                throw new DomainException(DomainErrorCode.NotFound, "Post not found.");
+
+            post.Content = request.Content;
+            post.Title = request.Title;
+            
             var response = await _blogRepository.UpdatePostAsync(post);
             return post.DtoToPost();
         }
